// api/models/driverCreateRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface DriverCreateRequest {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  broadcastName: string;
  /**
   * @minLength 3
   * @maxLength 3
   */
  acronym: string;
  /**
   * @minimum 1
   * @maximum 99
   */
  driverNumber: number;
  imageUrl?: string;
  /**
   * @minLength 2
   * @maxLength 2
   */
  countryCode: string;
}